+++ mbircone/src/MBIRModularUtilities3D.h
@@ -127,9 +127,9 @@ struct ImageParams
 {
     /* Location of the corner of the first voxel corresponding to
      (j_x, j_y, j_z) = (0, 0, 0). */
-    double x_0;
-    double y_0;
-    double z_0;
+    float x_0;
+    float y_0;
+    float z_0;
     
     /* Number of voxels in x, y, z direction. */
     long int N_x;
@@ -137,8 +137,8 @@ struct ImageParams
     long int N_z;
 
     /* Dimensions of a voxel */
-    double Delta_xy;
-    double Delta_z;
+    float Delta_xy;
+    float Delta_z;
     
     /**
      *      Region of Interest (roi) parameters
@@ -183,27 +183,27 @@ struct SinoParams
     long int N_dw;
 
     /* Detector width and spacing in v-direction and w-direction. */
-    double Delta_dv;
-    double Delta_dw;
+    float Delta_dv;
+    float Delta_dw;
 
     /* Number of discrete view angles. */
     long int N_beta;
     
     /* The source location on the u-axis. Assume u_s < 0. */
-    double u_s;
+    float u_s;
     
     /* The location (u_r, v_r) of the center of rotation. */
-    double u_r;
-    double v_r;
+    float u_r;
+    float v_r;
     
     /* The location (u,v,w) of the corner of the first detector corresponding to
      (i_v, i_w) = (0, 0). All points on the detector have u = u_d0. */
-    double u_d0;
-    double v_d0;
-    double w_d0;
+    float u_d0;
+    float v_d0;
+    float w_d0;
     
     /* Noise variance estimation */
-    double weightScaler_value;       /* Weight_true = Weight / weightScaler_value */
+    float weightScaler_value;       /* Weight_true = Weight / weightScaler_value */
 };
 
 
@@ -222,7 +222,7 @@ struct Sino
 struct ViewAngleList
 {
     long int N_beta;
-    double *beta;
+    float *beta;
 };
 
 
@@ -232,31 +232,31 @@ struct ReconParams
     /**
      *     Prior
      */
-    double priorWeight_QGGMRF;                  /* Prior mode: (0: off, 1: QGGMRF, 2: proximal mapping) */
-    double priorWeight_proxMap;                  /* Prior mode: (0: off, 1: QGGMRF, 2: proximal mapping) */
+    float priorWeight_QGGMRF;                  /* Prior mode: (0: off, 1: QGGMRF, 2: proximal mapping) */
+    float priorWeight_proxMap;                  /* Prior mode: (0: off, 1: QGGMRF, 2: proximal mapping) */
     
     /* QGGMRF */
-        double q;                   /* q: QGGMRF parameter (q>1, typical choice q=2) */
-        double p;                   /* p: QGGMRF parameter (1<=p<q) */
-        double T;                   /* T: QGGMRF parameter */
-        double sigmaX;              /* sigmaX: QGGMRF parameter */
-        double bFace;               /* bFace: relative neighbor weight: cube faces */
-        double bEdge;               /* bEdge: relative neighbor weight: cube edges */
-        double bVertex;             /* bVertex: relative neighbor weight: cube vertices */
+        float q;                   /* q: QGGMRF parameter (q>1, typical choice q=2) */
+        float p;                   /* p: QGGMRF parameter (1<=p<q) */
+        float T;                   /* T: QGGMRF parameter */
+        float sigmaX;              /* sigmaX: QGGMRF parameter */
+        float bFace;               /* bFace: relative neighbor weight: cube faces */
+        float bEdge;               /* bEdge: relative neighbor weight: cube edges */
+        float bVertex;             /* bVertex: relative neighbor weight: cube vertices */
     /* Proximal Mapping */
-        double sigma_lambda;        /* sigma_lambda: Proximal mapping scalar */
+        float sigma_lambda;        /* sigma_lambda: Proximal mapping scalar */
         int is_positivity_constraint;
     
     /**
      *      Stopping Conditions
      */
-    double stopThresholdChange_pct;           /* stop threshold (%) */
-    double stopThesholdRWFE_pct;
-    double stopThesholdRUFE_pct;
+    float stopThresholdChange_pct;           /* stop threshold (%) */
+    float stopThesholdRWFE_pct;
+    float stopThesholdRUFE_pct;
     int MaxIterations;              /* maximum number of iterations */
     char relativeChangeMode[200];
-    double relativeChangeScaler;
-    double relativeChangePercentile;
+    float relativeChangeScaler;
+    float relativeChangePercentile;
 
 
     /**
@@ -273,14 +273,14 @@ struct ReconParams
      */
     char weightScaler_estimateMode[200];     /* Estimate weight scaler? 1: Yes. 0: Use user specified value */
     char weightScaler_domain[200];     
-    double weightScaler_value;            /* User specified weight scaler */
+    float weightScaler_value;            /* User specified weight scaler */
 
 
     /* NHICD stuff */
     char NHICD_Mode[200];
-    double NHICD_ThresholdAllVoxels_ErrorPercent;
-    double NHICD_percentage;
-    double NHICD_random;
+    float NHICD_ThresholdAllVoxels_ErrorPercent;
+    float NHICD_percentage;
+    float NHICD_random;
 
     /* Misc */
     int verbosity;
@@ -293,14 +293,14 @@ struct SysMatrix
     long int i_vstride_max;        /* max_{i,j}(i_vstride) */
     long int i_wstride_max;        /* max_{i,j}(i_wstride) */
     long int N_u;
-    double B_ij_max;
-    double C_ij_max;
-    double B_ij_scaler;  /* B_ij_true = B_ij * B_ij_scaler*/
-    double C_ij_scaler;  /* C_ij_true = C_ij * C_ij_scaler*/
+    float B_ij_max;
+    float C_ij_max;
+    float B_ij_scaler;  /* B_ij_true = B_ij * B_ij_scaler*/
+    float C_ij_scaler;  /* C_ij_true = C_ij * C_ij_scaler*/
 
-    double Delta_u;     /* = Delta_xy / rho      */
-    double u_0;
-    double u_1;
+    float Delta_u;     /* = Delta_xy / rho      */
+    float u_0;
+    float u_1;
 
     BIJDATATYPE ***B;                   /* [N_x][N_y][N_beta*i_vstride_max]  */
     INDEXSTARTSTOPDATATYPE ***i_vstart; /* [N_x][N_y][N_beta]           */
@@ -326,29 +326,29 @@ struct ICDInfo3DCone
 
     float neighborsFace[6], neighborsEdge[12], neighborsVertex[8];
     float lastChange;
-    double old_xj; /* current pixel value */
-    double wghtRecon_j;
+    float old_xj; /* current pixel value */
+    float wghtRecon_j;
 
     /**
      *      The following is arbitrary when "ICDStep3DCone" is called
      *      and will be used as as variables inside "ICDStep3DCone".
      */
-    double Delta_xj;          /* Delta_xj = x_j^new - x_j^old */
-    double proxMapInput_j;    /* jth voxel of the proxMapInput */
+    float Delta_xj;          /* Delta_xj = x_j^new - x_j^old */
+    float proxMapInput_j;    /* jth voxel of the proxMapInput */
 
-    double theta1_f;
-    double theta2_f;
-    double theta1_p_QGGMRF;
-    double theta2_p_QGGMRF;
-    double theta1_p_proxMap;
-    double theta2_p_proxMap;
+    float theta1_f;
+    float theta2_f;
+    float theta1_p_QGGMRF;
+    float theta2_p_QGGMRF;
+    float theta1_p_proxMap;
+    float theta2_p_proxMap;
 
 };
 
 struct PartialTheta
 {
-    double t1;
-    double t2;
+    float t1;
+    float t2;
 };
 
 struct ParallelAux
@@ -358,28 +358,28 @@ struct ParallelAux
     struct PartialTheta **partialTheta;     /* [numThreads][N_M_max] */
     long int *j_u;
     long int *i_v;
-    double *B_ij;
+    float *B_ij;
     long int *k_M;
     long int *j_z;
     long int *i_w;
-    double *A_ij;
+    float *A_ij;
 };
 
 struct SpeedAuxICD
 {
     long int numberUpdatedVoxels;
-    double tic;
-    double toc;
-    double voxelsPerSecond;    
+    float tic;
+    float toc;
+    float voxelsPerSecond;    
 };
 
 struct IterationStatistics
 {
-    double cost;
-    double relUpdate;
-    double weightScaler_value;
-    double voxelsPerSecond;
-    double ticToc_iteration;
+    float cost;
+    float relUpdate;
+    float weightScaler_value;
+    float voxelsPerSecond;
+    float ticToc_iteration;
 };
 
 
@@ -387,18 +387,18 @@ struct ReconAux
 { 
     int NHICD_isPartialUpdateActive;
     long int *NHICD_numUpdatedVoxels;
-    double *NHICD_totalValueChange;
+    float *NHICD_totalValueChange;
     int *NHICD_isPartialZiplineHot;
     
-    double lastChangeThreshold;
+    float lastChangeThreshold;
     int N_M_max;
-    double totalEquits;
+    float totalEquits;
 
-    double relativeWeightedForwardError;
-    double relativeUnweightedForwardError;
+    float relativeWeightedForwardError;
+    float relativeUnweightedForwardError;
 
-    double TotalValueChange;
-    double TotalVoxelValue;
+    float TotalValueChange;
+    float TotalVoxelValue;
     long int NumUpdatedVoxels;
 
     float NHICD_neighborFilter[3][3];
@@ -412,17 +412,17 @@ void backProjectlike3DCone( float ***x_out, float ***y_in, struct ImageParams *i
 
 void initializeWghtRecon(struct SysMatrix *A, struct Sino *sino, struct Image *img, struct ReconParams *reconParams);
 
-double computeAvgWghtRecon(struct Image *img);
+float computeAvgWghtRecon(struct Image *img);
     
 void computeSecondaryReconParams(struct ReconParams *reconParams, struct ImageParams *imgParams);
 
-void invertDoubleMatrix(double **A, double ** A_inv, int size);
+void invertDoubleMatrix(float **A, float ** A_inv, int size);
 
-double computeNormSquaredFloatArray(float *arr, long int len);
+float computeNormSquaredFloatArray(float *arr, long int len);
 
-double computeRelativeRMSEFloatArray(float *arr1, float *arr2, long int len);
+float computeRelativeRMSEFloatArray(float *arr1, float *arr2, long int len);
 
-double computeSinogramWeightedNormSquared(struct Sino *sino, float ***arr);
+float computeSinogramWeightedNormSquared(struct Sino *sino, float ***arr);
 
 char isInsideMask(long int i_1, long int i_2, long int N1, long int N2);
 
@@ -432,7 +432,7 @@ void copyImage2ROI(struct Image *img);
 
 void applyMask(float ***arr, long int N1, long int N2, long int N3);
 
-void floatArray_z_equals_aX_plus_bY(float *Z, double a, float *X, double b, float *Y, long int len);
+void floatArray_z_equals_aX_plus_bY(float *Z, float a, float *X, float b, float *Y, long int len);
 
 void setFloatArray2Value(float *arr, long int len, float value);
 
@@ -472,26 +472,26 @@ void shuffleIntArray(int *arr, long int len);
 
 void shuffleLongIntArray(long int *arr, long int len);
 
-int bernoulli(double p);
+int bernoulli(float p);
 
 long int uniformIntegerRV(long int l, long int h);
 
-long int almostUniformIntegerRV(double mean, int sigma);
+long int almostUniformIntegerRV(float mean, int sigma);
 
 
 
 /**************************************** tic toc ****************************************/
-void tic(double *ticToc);
+void tic(float *ticToc);
 
-void toc(double *ticToc);
+void toc(float *ticToc);
 
-void ticTocDisp(double ticToc, char *ticTocName);
+void ticTocDisp(float ticToc, char *ticTocName);
 
 /**************************************** timer ****************************************/
 
-void timer_reset(double *timer);
+void timer_reset(float *timer);
 
-int timer_hasPassed(double *timer, double time_passed);
+int timer_hasPassed(float *timer, float time_passed);
 
 
 /**************************************** percentile stuff ****************************************/
